Doc XML: java.awt.PrinterJob is returned by the pdf_output plugin and used as parameter type by the controller.print...() functions and as such shown up in the Wiki docs like java.awt.PrinterJob
i just added a PrinterJob class in the pdf_output that is servoy documented. 
Don't know if this helps because we can't have java.awt.PrinterJob documented because we can't annotate that class. Also we can't just create a subclass of PrinterJob in servoy's code base (and use that as the input of controller.print) because the pdf printer is not sub class of that printerjob then.
Case: SVY-2556